// web/src/app/api/generated/src/models/Dose.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Motion Ecosystem API
 * API for participant and specialist workflows
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { AerobicDose } from './AerobicDose';
import {
  instanceOfAerobicDose,
  AerobicDoseFromJSON,
  AerobicDoseFromJSONTyped,
  AerobicDoseToJSON,
} from './AerobicDose';
import type { BreathingDose } from './BreathingDose';
import {
  instanceOfBreathingDose,
  BreathingDoseFromJSON,
  BreathingDoseFromJSONTyped,
  BreathingDoseToJSON,
} from './BreathingDose';
import type { IsometricDose } from './IsometricDose';
import {
  instanceOfIsometricDose,
  IsometricDoseFromJSON,
  IsometricDoseFromJSONTyped,
  IsometricDoseToJSON,
} from './IsometricDose';
import type { MobilityDose } from './MobilityDose';
import {
  instanceOfMobilityDose,
  MobilityDoseFromJSON,
  MobilityDoseFromJSONTyped,
  MobilityDoseToJSON,
} from './MobilityDose';
import type { StrengthDose } from './StrengthDose';
import {
  instanceOfStrengthDose,
  StrengthDoseFromJSON,
  StrengthDoseFromJSONTyped,
  StrengthDoseToJSON,
} from './StrengthDose';
import type { StretchDose } from './StretchDose';
import {
  instanceOfStretchDose,
  StretchDoseFromJSON,
  StretchDoseFromJSONTyped,
  StretchDoseToJSON,
} from './StretchDose';

/**
 * @type Dose
 *
 * @export
 */
export type Dose =
  | ({ type: 'AEROBIC' } & AerobicDose)
  | ({ type: 'BREATHING' } & BreathingDose)
  | ({ type: 'ISOMETRIC' } & IsometricDose)
  | ({ type: 'MOBILITY' } & MobilityDose)
  | ({ type: 'STRENGTH' } & StrengthDose)
  | ({ type: 'STRETCH' } & StretchDose);

export function DoseFromJSON(json: any): Dose {
  return DoseFromJSONTyped(json, false);
}

export function DoseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Dose {
  if (json == null) {
    return json;
  }
  switch (json['type']) {
    case 'AEROBIC':
      return Object.assign({}, AerobicDoseFromJSONTyped(json, true), { type: 'AEROBIC' } as const);
    case 'BREATHING':
      return Object.assign({}, BreathingDoseFromJSONTyped(json, true), {
        type: 'BREATHING',
      } as const);
    case 'ISOMETRIC':
      return Object.assign({}, IsometricDoseFromJSONTyped(json, true), {
        type: 'ISOMETRIC',
      } as const);
    case 'MOBILITY':
      return Object.assign({}, MobilityDoseFromJSONTyped(json, true), {
        type: 'MOBILITY',
      } as const);
    case 'STRENGTH':
      return Object.assign({}, StrengthDoseFromJSONTyped(json, true), {
        type: 'STRENGTH',
      } as const);
    case 'STRETCH':
      return Object.assign({}, StretchDoseFromJSONTyped(json, true), { type: 'STRETCH' } as const);
    default:
      return json;
  }
}

export function DoseToJSON(json: any): any {
  return DoseToJSONTyped(json, false);
}

export function DoseToJSONTyped(value?: Dose | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }
  switch (value['type']) {
    case 'AEROBIC':
      return Object.assign({}, AerobicDoseToJSON(value), { type: 'AEROBIC' } as const);
    case 'BREATHING':
      return Object.assign({}, BreathingDoseToJSON(value), { type: 'BREATHING' } as const);
    case 'ISOMETRIC':
      return Object.assign({}, IsometricDoseToJSON(value), { type: 'ISOMETRIC' } as const);
    case 'MOBILITY':
      return Object.assign({}, MobilityDoseToJSON(value), { type: 'MOBILITY' } as const);
    case 'STRENGTH':
      return Object.assign({}, StrengthDoseToJSON(value), { type: 'STRENGTH' } as const);
    case 'STRETCH':
      return Object.assign({}, StretchDoseToJSON(value), { type: 'STRETCH' } as const);
    default:
      return value;
  }
}